requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-filter
drf-spectacular
psycopg2-binary
pillow
celery
redis
django-environ
django-cors-headers
django-redis
django-celery-results
django-celery-beat
